#48: Move adapters to own folder
