+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/adapters/ClaimListAdapter.java
@@ -23,7 +23,7 @@
 
 // Sourced (Mar 8, 2015)
 //   http://www.piwai.info/android-adapter-good-practices/
-package ca.ualberta.cs.shinyexpensetracker;
+package ca.ualberta.cs.shinyexpensetracker.adapters;
 
 import java.text.SimpleDateFormat;
 import java.util.Locale;
@@ -34,6 +34,11 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
+import ca.ualberta.cs.shinyexpensetracker.Application;
+import ca.ualberta.cs.shinyexpensetracker.ExpenseClaimController;
+import ca.ualberta.cs.shinyexpensetracker.R;
+import ca.ualberta.cs.shinyexpensetracker.R.id;
+import ca.ualberta.cs.shinyexpensetracker.R.layout;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 
 public class ClaimListAdapter extends BaseAdapter {
