+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/adapters/SectionsPagerAdapter.java
@@ -2,7 +2,7 @@
  * Provides views to tab fragments
  */
 
-package ca.ualberta.cs.shinyexpensetracker;
+package ca.ualberta.cs.shinyexpensetracker.adapters;
 
 import java.util.Locale;
 
@@ -12,6 +12,8 @@ import android.content.Context;
 import android.support.v13.app.FragmentPagerAdapter;
 import android.util.SparseArray;
 import android.view.ViewGroup;
+import ca.ualberta.cs.shinyexpensetracker.R;
+import ca.ualberta.cs.shinyexpensetracker.R.string;
 import ca.ualberta.cs.shinyexpensetracker.fragments.ClaimSummaryFragment;
 import ca.ualberta.cs.shinyexpensetracker.fragments.DestinationListFragment;
 import ca.ualberta.cs.shinyexpensetracker.fragments.ExpenseItemListFragment;
