+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/fragments/ClaimSummaryFragment.java
@@ -12,9 +12,9 @@ import android.widget.ListView;
 import android.widget.TextView;
 import ca.ualberta.cs.shinyexpensetracker.Application;
 import ca.ualberta.cs.shinyexpensetracker.ExpenseClaimController;
-import ca.ualberta.cs.shinyexpensetracker.ExpenseTotalsAdapter;
 import ca.ualberta.cs.shinyexpensetracker.IView;
 import ca.ualberta.cs.shinyexpensetracker.R;
+import ca.ualberta.cs.shinyexpensetracker.adapters.ExpenseTotalsAdapter;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 
 /**
