+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/fragments/DestinationListFragment.java
@@ -14,11 +14,11 @@ import android.widget.AdapterView.OnItemLongClickListener;
 import android.widget.ListView;
 import android.widget.TextView;
 import ca.ualberta.cs.shinyexpensetracker.Application;
-import ca.ualberta.cs.shinyexpensetracker.DestinationItemAdapter;
 import ca.ualberta.cs.shinyexpensetracker.ExpenseClaimController;
 import ca.ualberta.cs.shinyexpensetracker.IView;
 import ca.ualberta.cs.shinyexpensetracker.R;
 import ca.ualberta.cs.shinyexpensetracker.activities.AddDestinationActivity;
+import ca.ualberta.cs.shinyexpensetracker.adapters.DestinationItemAdapter;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 
 /**
