+++ ShinyExpenseTrackerTests/src/ca/ualberta/cs/shinyexpensetracker/test/ExpenseTotalsAdapterTest.java
@@ -6,7 +6,7 @@ import java.util.Date;
 import android.test.AndroidTestCase;
 import android.view.View;
 import android.widget.TextView;
-import ca.ualberta.cs.shinyexpensetracker.ExpenseTotalsAdapter;
+import ca.ualberta.cs.shinyexpensetracker.adapters.ExpenseTotalsAdapter;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseItem;
 import ca.ualberta.cs.shinyexpensetracker.models.ExpenseItem.Category;
